requirements.txt
requests
markdown2
atlassian-python-api
argparse
html2text
pyyaml